+++ packages/analytics-react-native-test/test/autocapture/element-interactions.harness.tsx
@@ -35,12 +35,11 @@ function ButtonHarness() {
             console.log('Button pressed');
           },
           {
-            event: 'Press',
             testID: 'my-button',
             component: 'ButtonHarness',
             element: 'Button',
             accessibilityLabel: 'Button accessibility label',
-            action: 'onPress',
+            action: 'press',
           },
         )}
       />
@@ -87,7 +86,7 @@ describe('autocapture.elementInteractions', () => {
       '[Amplitude] Target Component': 'ButtonHarness',
       '[Amplitude] Target Element': 'Button',
       '[Amplitude] Target Test ID': 'my-button',
-      '[Amplitude] Action': 'onPress',
+      '[Amplitude] Action': 'press',
     });
   });
 });
@@ -114,7 +113,7 @@ describe('autocapture.elementInteractions is undefined', () => {
     capture.clear();
   });
 
-  it('tracks onPress event on button', async () => {
+  it('onPress event is not tracked', async () => {
     render(<ButtonHarness />);
     const button = await screen.findByTestId('my-button');
     await userEvent.press(button);