+++ packages/analytics-react-native/src/amp-capture.ts
@@ -1,20 +1,9 @@
-export const EVENT_TYPE_VALUES = {
-  Press: 'Press',
-  LongPress: 'LongPress',
-  ValueChange: 'ValueChange',
-  ChangeText: 'ChangeText',
-  SubmitEditing: 'SubmitEditing',
-} as const;
-
-export type EVENT_TYPE = (typeof EVENT_TYPE_VALUES)[keyof typeof EVENT_TYPE_VALUES];
-
 export type AmpCaptureProperties = {
   action?: string;
   accessibilityLabel?: string;
   component?: string;
   element?: string;
   testID?: string;
-  event: EVENT_TYPE;
 };
 
 const callbacks: ((properties: AmpCaptureProperties) => void)[] = [];
@@ -22,9 +11,14 @@ const callbacks: ((properties: AmpCaptureProperties) => void)[] = [];
 export function subscribe(callback: (properties: AmpCaptureProperties) => void) {
   callbacks.push(callback);
   return () => {
-    callbacks.splice(callbacks.indexOf(callback), 1);
+    const index = callbacks.indexOf(callback);
+    if (index >= 0) {
+      callbacks.splice(index, 1);
+    }
   };
 }
+
+let isAmpCapturing = false;
 
 export function ampCapture<Args extends unknown[], Return>(
   func: (...args: Args) => Return,
@@ -34,11 +28,21 @@ export function ampCapture<Args extends unknown[], Return>(
     return func;
   }
   return (...args: Args) => {
-    try {
-      callbacks.forEach((callback) => callback(properties));
-    } catch (error) {
-      // swallow errors
+    if (!isAmpCapturing) {
+      // only call "callbacks" if not nested inside another ampCapture
+      try {
+        isAmpCapturing = true;
+        try {
+          callbacks.forEach((callback) => callback(properties));
+        } catch (error) {
+          // swallow errors
+        }
+        return func(...args);
+      } finally {
+        isAmpCapturing = false;
+      }
+    } else {
+      return func(...args);
     }
-    return func(...args);
   };
 }