+++ packages/plugin-experiment-browser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.38](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-experiment-browser@1.0.0-beta.37...@amplitude/plugin-experiment-browser@1.0.0-beta.38) (2026-07-22)
+
+**Note:** Version bump only for package @amplitude/plugin-experiment-browser
+
+
+
+
+
+# [1.0.0-beta.37](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-experiment-browser@1.0.0-beta.36...@amplitude/plugin-experiment-browser@1.0.0-beta.37) (2026-07-20)
+
+**Note:** Version bump only for package @amplitude/plugin-experiment-browser
+
+
+
+
+
 # [1.0.0-beta.36](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-experiment-browser@1.0.0-beta.35...@amplitude/plugin-experiment-browser@1.0.0-beta.36) (2026-07-17)
 
 **Note:** Version bump only for package @amplitude/plugin-experiment-browser