+++ packages/plugin-web-vitals-browser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.43](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-web-vitals-browser@1.1.42...@amplitude/plugin-web-vitals-browser@1.1.43) (2026-07-22)
+
+**Note:** Version bump only for package @amplitude/plugin-web-vitals-browser
+
+
+
+
+
+## [1.1.42](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-web-vitals-browser@1.1.41...@amplitude/plugin-web-vitals-browser@1.1.42) (2026-07-20)
+
+**Note:** Version bump only for package @amplitude/plugin-web-vitals-browser
+
+
+
+
+
 ## [1.1.41](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-web-vitals-browser@1.1.40...@amplitude/plugin-web-vitals-browser@1.1.41) (2026-07-17)
 
 **Note:** Version bump only for package @amplitude/plugin-web-vitals-browser